cmd/swarm, swarm: address PR comments
