+++ cmd/swarm/access.go
@@ -103,7 +103,7 @@ func accessNewPK(ctx *cli.Context) {
 		}
 	}
 }
-
+	
 func accessNewACT(ctx *cli.Context) {
 	args := ctx.Args()
 	if len(args) != 1 {
