+++ swarm/api/manifest.go
@@ -19,8 +19,6 @@ package api
 import (
 	"bytes"
 	"context"
-	"crypto/ecdsa"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -29,11 +27,7 @@ import (
 	"strings"
 	"time"
 
-	"golang.org/x/crypto/scrypt"
-
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/ethereum/go-ethereum/crypto/ecies"
 	"github.com/ethereum/go-ethereum/swarm/log"
 	"github.com/ethereum/go-ethereum/swarm/storage"
 )
@@ -60,162 +54,6 @@ type ManifestEntry struct {
 	ModTime     time.Time    `json:"mod_time,omitempty"`
 	Status      int          `json:"status,omitempty"`
 	Access      *AccessEntry `json:"access,omitempty"`
-}
-type AccessEntry struct {
-	Type      AccessType
-	Publisher string
-	Salt      []byte
-	Act       string
-	KdfParams *KdfParams
-}
-
-type DecryptFunc func(*ManifestEntry) error
-
-func (a *AccessEntry) MarshalJSON() (out []byte, err error) {
-
-	return json.Marshal(struct {
-		Type      AccessType `json:"type,omitempty"`
-		Publisher string     `json:"publisher,omitempty"`
-		Salt      string     `json:"salt,omitempty"`
-		Act       string     `json:"act,omitempty"`
-		KdfParams *KdfParams `json:"kdf_params,omitempty"`
-	}{
-		Type:      a.Type,
-		Publisher: a.Publisher,
-		Salt:      hex.EncodeToString(a.Salt),
-		Act:       a.Act,
-		KdfParams: a.KdfParams,
-	})
-
-}
-
-func (a *AccessEntry) UnmarshalJSON(value []byte) error {
-	v := struct {
-		Type      AccessType `json:"type,omitempty"`
-		Publisher string     `json:"publisher,omitempty"`
-		Salt      string     `json:"salt,omitempty"`
-		Act       string     `json:"act,omitempty"`
-		KdfParams *KdfParams `json:"kdf_params,omitempty"`
-	}{}
-
-	err := json.Unmarshal(value, &v)
-	if err != nil {
-		return err
-	}
-	a.Act = v.Act
-	a.KdfParams = v.KdfParams
-	a.Publisher = v.Publisher
-	a.Salt, err = hex.DecodeString(v.Salt)
-	if err != nil {
-		return err
-	}
-	if len(a.Salt) != 32 {
-		return errors.New("salt should be 32 bytes long")
-	}
-	a.Type = v.Type
-	return nil
-}
-
-type KdfParams struct {
-	N int `json:"n"`
-	P int `json:"p"`
-	R int `json:"r"`
-}
-
-type AccessType string
-
-const AccessTypePass = AccessType("pass")
-const AccessTypePK = AccessType("pk")
-const AccessTypeACT = AccessType("act")
-
-func NewAccessEntryPassword(salt []byte, kdfParams *KdfParams) (*AccessEntry, error) {
-	if len(salt) != 32 {
-		return nil, fmt.Errorf("salt should be 32 bytes long")
-	}
-	return &AccessEntry{
-		Type:      AccessTypePass,
-		Salt:      salt,
-		KdfParams: kdfParams,
-	}, nil
-}
-
-func NewAccessEntryPK(publisher string, salt []byte) (*AccessEntry, error) {
-	if len(publisher) != 66 {
-		return nil, fmt.Errorf("publisher should be 66 characters long, got %d", len(publisher))
-	}
-	if len(salt) != 32 {
-		return nil, fmt.Errorf("salt should be 32 bytes long")
-	}
-	return &AccessEntry{
-		Type:      AccessTypePK,
-		Publisher: publisher,
-		Salt:      salt,
-	}, nil
-}
-
-func NewAccessEntryACT(publisher string, salt []byte, act string) (*AccessEntry, error) {
-	if len(salt) != 32 {
-		return nil, fmt.Errorf("salt should be 32 bytes long")
-	}
-	if len(publisher) != 66 {
-		return nil, fmt.Errorf("publisher should be 66 characters long")
-	}
-
-	return &AccessEntry{
-		Type:      AccessTypeACT,
-		Publisher: publisher,
-		Salt:      salt,
-		Act:       act,
-	}, nil
-}
-
-func NOOPDecrypt(*ManifestEntry) error {
-	return nil
-}
-
-var DefaultKdfParams = NewKdfParams(262144, 1, 8)
-
-func NewKdfParams(n, p, r int) *KdfParams {
-
-	return &KdfParams{
-		N: n,
-		P: p,
-		R: r,
-	}
-}
-
-// NewSessionKeyPassword creates a session key based on a shared secret (password) and the given salt
-// and kdf parameters in the access entry
-func NewSessionKeyPassword(password string, accessEntry *AccessEntry) ([]byte, error) {
-	if accessEntry.Type != AccessTypePass {
-		return nil, errors.New("incorrect access entry type")
-	}
-	return scrypt.Key(
-		[]byte(password),
-		accessEntry.Salt,
-		accessEntry.KdfParams.N,
-		accessEntry.KdfParams.R,
-		accessEntry.KdfParams.P,
-		32,
-	)
-}
-
-// NewSessionKeyPK creates a new ACT Session Key using an ECDH shared secret for the given key pair and the given salt value
-func NewSessionKeyPK(private *ecdsa.PrivateKey, public *ecdsa.PublicKey, salt []byte) ([]byte, error) {
-	granteePubEcies := ecies.ImportECDSAPublic(public)
-	privateKey := ecies.ImportECDSA(private)
-
-	bytes, err := privateKey.GenerateShared(granteePubEcies, 16, 16)
-	if err != nil {
-		return nil, err
-	}
-	bytes = append(salt, bytes...)
-	sessionKey := crypto.Keccak256(bytes)
-	return sessionKey, nil
-}
-
-func (a *API) NodeSessionKey(privateKey *ecdsa.PrivateKey, publicKey *ecdsa.PublicKey, salt []byte) ([]byte, error) {
-	return NewSessionKeyPK(privateKey, publicKey, salt)
 }
 
 // ManifestList represents the result of listing files in a manifest
